modelo terminado primera version

+++ SimulatedAnneling/Model/SimulatedAnneling/ISolution.cs
@@ -18,7 +18,7 @@ namespace SimulatedAnneling.Model.SimulatedAnneling
         ///<returns>
         ///Función de costo asociada a la solución
         ///</returns>
-        float calculateCostFunction();
+        double calculateCostFunction();
 
         ///<summary>
         ///Obtiene un vecino a la solución actual a partir de un elemento random
